[CSL-1621] Move command execution to Command in Rubbish
